backend/data-type/EU-bank/clustering/Email/requirements.txt
spacy==3.7.2
pydantic==2.5.0
gensim==4.3.2
hdbscan==0.8.33
numpy>=1.26.0
pandas>=2.1.0
umap-learn>=0.5.5
scikit-learn>=1.3.0
pymongo==4.6.0
python-dotenv==1.0.0
